# docs/api/openapi-rest.json
{
  "components": {
    "schemas": {
      "DiffResponseDict": {
        "properties": {
          "diff": {
            "title": "Diff",
            "type": "string"
          },
          "from_date": {
            "title": "From Date",
            "type": "string"
          },
          "law_id": {
            "title": "Law Id",
            "type": "string"
          },
          "to_date": {
            "title": "To Date",
            "type": "string"
          }
        },
        "required": [
          "law_id",
          "from_date",
          "to_date",
          "diff"
        ],
        "title": "DiffResponseDict",
        "type": "object"
      },
      "ErrorResponseDict": {
        "description": "D-052 error envelope shape, for OpenAPI documentation only (PR\nreview fix #3). Every REST error body carries at least `code`; the\nremaining keys vary per code — see docs/api/error-codes.md for the\nfull per-code payload contracts.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "ErrorResponseDict",
        "type": "object"
      },
      "GetArticleResponseDict": {
        "properties": {
          "article": {
            "title": "Article",
            "type": "string"
          },
          "commit_hash": {
            "title": "Commit Hash",
            "type": "string"
          },
          "implicit": {
            "title": "Implicit",
            "type": "boolean"
          },
          "law_id": {
            "title": "Law Id",
            "type": "string"
          },
          "paragraph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "text_hash": {
            "title": "Text Hash",
            "type": "string"
          },
          "warnings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "law_id",
          "article",
          "paragraph",
          "text",
          "text_hash",
          "commit_hash",
          "implicit",
          "warnings"
        ],
        "title": "GetArticleResponseDict",
        "type": "object"
      },
      "GetLawResponseDict": {
        "properties": {
          "amendment_history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Amendment History",
            "type": "array"
          },
          "body_markdown": {
            "title": "Body Markdown",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "commit_hash": {
            "title": "Commit Hash",
            "type": "string"
          },
          "dv_issue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dv Issue"
          },
          "dv_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dv Year"
          },
          "effective_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Date"
          },
          "eli": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eli"
          },
          "fecha_publicacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Publicacion"
          },
          "identificador": {
            "title": "Identificador",
            "type": "string"
          },
          "law_id": {
            "title": "Law Id",
            "type": "string"
          },
          "titulo": {
            "title": "Titulo",
            "type": "string"
          },
          "ultima_actualizacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ultima Actualizacion"
          },
          "warnings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "law_id",
          "identificador",
          "titulo",
          "category",
          "fecha_publicacion",
          "ultima_actualizacion",
          "dv_issue",
          "dv_year",
          "effective_date",
          "eli",
          "amendment_history",
          "commit_hash",
          "body_markdown",
          "warnings"
        ],
        "title": "GetLawResponseDict",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LawListResponseDict": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LawSummaryDict"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "items"
        ],
        "title": "LawListResponseDict",
        "type": "object"
      },
      "LawSummaryDict": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "first_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Version"
          },
          "identificador": {
            "title": "Identificador",
            "type": "string"
          },
          "latest_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Version"
          },
          "law_id": {
            "title": "Law Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version_count": {
            "title": "Version Count",
            "type": "integer"
          }
        },
        "required": [
          "law_id",
          "identificador",
          "title",
          "category",
          "status",
          "first_version",
          "latest_version",
          "version_count"
        ],
        "title": "LawSummaryDict",
        "type": "object"
      },
      "MatchedSegmentDict": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "label"
        ],
        "title": "MatchedSegmentDict",
        "type": "object"
      },
      "SearchHitDict": {
        "properties": {
          "body_snippet": {
            "title": "Body Snippet",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "identificador": {
            "title": "Identificador",
            "type": "string"
          },
          "law_id": {
            "title": "Law Id",
            "type": "string"
          },
          "matched": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchedSegmentDict"
              },
              {
                "type": "null"
              }
            ]
          },
          "relevance": {
            "title": "Relevance",
            "type": "number"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "title_snippet": {
            "title": "Title Snippet",
            "type": "string"
          }
        },
        "required": [
          "law_id",
          "identificador",
          "title",
          "category",
          "title_snippet",
          "body_snippet",
          "matched",
          "relevance"
        ],
        "title": "SearchHitDict",
        "type": "object"
      },
      "StatsResponseDict": {
        "properties": {
          "by_category": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Category",
            "type": "object"
          },
          "by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Status",
            "type": "object"
          },
          "latest_version_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Version Date"
          },
          "multi_version_acts": {
            "title": "Multi Version Acts",
            "type": "integer"
          },
          "total_acts": {
            "title": "Total Acts",
            "type": "integer"
          }
        },
        "required": [
          "total_acts",
          "by_category",
          "by_status",
          "multi_version_acts",
          "latest_version_date"
        ],
        "title": "StatsResponseDict",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionEntryDict": {
        "properties": {
          "commit_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Hash"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "dv_issue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dv Issue"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          }
        },
        "required": [
          "date",
          "dv_issue",
          "operation",
          "commit_hash"
        ],
        "title": "VersionEntryDict",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "legalize-bg REST API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/laws": {
      "get": {
        "operationId": "list_laws_api_v1_laws_get",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Estado"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LawListResponseDict"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`INDEX_MISSING` — catalog.db is missing tables/columns or unreadable at query time"
          }
        },
        "summary": "List Laws"
      }
    },
    "/api/v1/laws/{slug}": {
      "get": {
        "operationId": "get_law_api_v1_laws__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLawResponseDict"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`INVALID_DATE` — a `date`-like parameter is missing or not a valid YYYY-MM-DD string"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`LAW_NOT_FOUND` — no act matches the given slug, identificador, or title; `NO_VERSION_AT_DATE` — the requested date precedes the act's earliest recorded version, or it has none"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`AMBIGUOUS_NAME` — multiple acts share the given title"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`INDEX_STALE` — the catalog and corpus have diverged for this act; `INDEX_MISSING` — catalog.db is missing tables/columns or unreadable at query time"
          }
        },
        "summary": "Get Law"
      }
    },
    "/api/v1/laws/{slug}/articles/{art}": {
      "get": {
        "operationId": "get_article_api_v1_laws__slug__articles__art__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "art",
            "required": true,
            "schema": {
              "title": "Art",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetArticleResponseDict"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`INVALID_DATE` — a `date`-like parameter is missing or not a valid YYYY-MM-DD string; `INVALID_ARTICLE_SPEC` — the article spec could not be parsed, or a range (`чл. 5-7`) was given where only a single article is accepted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`LAW_NOT_FOUND` — no act matches the given slug, identificador, or title; `NO_VERSION_AT_DATE` — the requested date precedes the act's earliest recorded version, or it has none; `ARTICLE_NOT_FOUND` — the article/alinea spec parsed but no provisions row matches it"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`AMBIGUOUS_NAME` — multiple acts share the given title"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`INDEX_MISSING` — catalog.db is missing tables/columns or unreadable at query time"
          }
        },
        "summary": "Get Article"
      }
    },
    "/api/v1/laws/{slug}/diff": {
      "get": {
        "operationId": "diff_api_v1_laws__slug__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "title": "To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffResponseDict"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`INVALID_DATE` — a `date`-like parameter is missing or not a valid YYYY-MM-DD string; `INVALID_DATE_RANGE` — the `from` date is later than the `to` date"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`LAW_NOT_FOUND` — no act matches the given slug, identificador, or title; `NO_VERSION_AT_DATE` — the requested date precedes the act's earliest recorded version, or it has none"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`AMBIGUOUS_NAME` — multiple acts share the given title"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`DIFF_FAILED` — the underlying `git diff` invocation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`INDEX_MISSING` — catalog.db is missing tables/columns or unreadable at query time"
          }
        },
        "summary": "Diff"
      }
    },
    "/api/v1/laws/{slug}/history": {
      "get": {
        "operationId": "history_api_v1_laws__slug__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VersionEntryDict"
                  },
                  "title": "Response History Api V1 Laws  Slug  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`LAW_NOT_FOUND` — no act matches the given slug, identificador, or title"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`AMBIGUOUS_NAME` — multiple acts share the given title"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`INDEX_MISSING` — catalog.db is missing tables/columns or unreadable at query time"
          }
        },
        "summary": "History"
      }
    },
    "/api/v1/metrics": {
      "get": {
        "operationId": "get_metrics_api_v1_metrics_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Metrics Api V1 Metrics Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Metrics"
      }
    },
    "/api/v1/search": {
      "get": {
        "operationId": "search_api_v1_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "include_body",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Body",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchHitDict"
                  },
                  "title": "Response Search Api V1 Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`QUERY_TOO_BROAD` — the search query reduces to a bare category stop-word (e.g. `наредба`) and would match thousands of acts"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`INDEX_MISSING` — catalog.db is missing tables/columns or unreadable at query time"
          }
        },
        "summary": "Search"
      }
    },
    "/api/v1/stats": {
      "get": {
        "operationId": "stats_api_v1_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponseDict"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDict"
                }
              }
            },
            "description": "`INDEX_MISSING` — catalog.db is missing tables/columns or unreadable at query time"
          }
        },
        "summary": "Stats"
      }
    },
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Healthz Healthz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz"
      }
    }
  }
}